+++ src/main/java/com/invadermonky/magicultureintegrations/config/MIConfigIntegrations.java
@@ -2,6 +2,7 @@ package com.invadermonky.magicultureintegrations.config;
 
 import com.invadermonky.magicultureintegrations.MagicultureIntegrations;
 import com.invadermonky.magicultureintegrations.config.generics.ConfigHeatables;
+import com.invadermonky.magicultureintegrations.config.tags.ModTags;
 import net.minecraftforge.common.config.Config;
 import net.minecraftforge.common.config.ConfigManager;
 import net.minecraftforge.fml.client.event.ConfigChangedEvent;
@@ -406,6 +407,7 @@ public class MIConfigIntegrations {
         public static void onConfigChanged(ConfigChangedEvent.OnConfigChangedEvent event) {
             if(event.getModID().equals(MagicultureIntegrations.MOD_ID)) {
                 ConfigManager.sync(MagicultureIntegrations.MOD_ID, Config.Type.INSTANCE);
+                ModTags.syncConfig();
             }
         }
     }